cling-tidy_update

+++ tasks/kulik_a_the_most_different_adjacent/seq/src/ops_seq.cpp
@@ -1,10 +1,10 @@
 #include "kulik_a_the_most_different_adjacent/seq/include/ops_seq.hpp"
 
-#include <numeric>
+#include <cmath>
+#include <cstddef>
 #include <vector>
 
 #include "kulik_a_the_most_different_adjacent/common/include/common.hpp"
-#include "util/include/util.hpp"
 
 namespace kulik_a_the_most_different_adjacent {
 
@@ -26,15 +26,15 @@ bool KulikATheMostDifferentAdjacentSEQ::RunImpl() {
   const auto n = input.size();
   OutType &ans = GetOutput();
   double mx = 0.;
-  int ind = 0;
+  size_t ind = 0;
   for (size_t i = 1; i < n; ++i) {
     if (std::abs(input[i - 1] - input[i]) > mx) {
       mx = std::abs(input[i - 1] - input[i]);
       ind = i - 1;
     }
   }
-  ans.first = ind;
-  ans.second = ind + 1;
+  ans.first = static_cast<int>(ind);
+  ans.second = static_cast<int>(ind + 1);
   return true;
 }
